perf/simplify batch 1: audio-thread caches, decode-path allocs, dead code
Review findings, lowest-risk batch:
- Ninjam::process(): cache the peak-meter exp() release coefficient and the
  samples-per-beat division (recomputed only when rate/tempo change) — was a
  transcendental per sample frame on the audio thread.
- AacDecoder: reuse the decode output buffer across packet callbacks (the one
  recurring allocation in the decode->ring path).
- NjAudio::decodeOgg(): decode straight into the destination vector (pre-sized
  from the known interval length) instead of chunk-bounce + insert; use the
  server's DownloadBegin.estSize to reserve interval buffers (capped).
- NjClient: collapse sendChat/sendAdmin/sendPrivate into sendChatParts; make
  subMask a run() local; drop redundant lastSend resets (keepalive is timer-
  pure); delete dead `authed`.
- Ninjam.cpp: delete TabField + NjPasswordField (Rack's ui::TextField already
  has nextField/prevField Tab cycling and a password mode); extract
  teardownJoin() shared by stopAll()/handleServerDrop().
- Stream.cpp: delete dead ReadCtx::self; move leftover body bytes instead of
  copying; hoist the duplicated "Stream ended" epilogue.
- Hls: drop looksLikeHls()'s dead body-sniff parameter.
- StationImporter::Probe now carries StreamClient::State instead of magic ints.
- NjAudio: delete dead Transfer::fourcc.

+++ src/Radio.cpp
@@ -171,7 +171,7 @@ struct Radio : Module {
 		setImportMsg("Auditioning\xe2\x80\xa6", false);
 		importer.start(u, [this]() {
 			akaudio::StationImporter::Probe p;
-			p.state = (int) stream.getState();
+			p.state = stream.getState();
 			p.frames = stream.producedFrames();
 			p.status = stream.getStatusText();
 			return p;
